Change .gitignore to work with build environment
I changed .gitignore becuase I switched from using make to using cmake.
So now I have all the build stuff in a directory that's not tracked by
git. This includes the compiled binary. I also changed it to work with
the YouCompleteMe plugin for vim, which needed the compile_commands.json
file to know what suggestions to give while typing.