Add implementation for part of the generate() method
I added implementation for the melody part of the generator. I needed
two different random number generators, one for the big picture eg.
whether to put a note/hold/rest. And one for choosing which note to put
in (A, B, C, etc.). There were changes made to CMakeLists.txt because I
needed to compile with c++11 support.