Finish implementation of generator in the menu
I added menu options for customizing all aspects of the song. I had to
change some stuff in Generator.cpp and .h because I realized that it
would be easier if the setKey() method took an int as an argument as
opposed to a char. This was because it would be a pain to do stuff with
multiple characters if the argument was a double-digit number.