Start using maps instead of an algorithm
I figured that I should use a bunch of std::maps instead of coming up
with a big complicated algorithm that would have bugs and poor
readability. The maps will be a good bit of typing, but I think it will
take less time to implement than the algorithm would have. Also the maps
will be much easier to make changes to than an algorithm with a mess of
if statements and for loops