Change from using cmake
I had a really ugly script in my vimrc that handled using cmake in my
project. I decided to go back to using good 'ol make. This means,
however, that I had to do something about compile_commands.json. I
didn't have it anymore with make, so I decided to use YCM's alternative
which is a .ycm_extra_conf.py file which was a nightmare to get working
in make, because I didn't really know how to use it that well. More
people will have make than cmake installed so that was another reason to
do it.