Get program producing sound
It finally makes a valid .wav file on it's own. I had to change the way
that it kept track of how long the file is, also I was writing WAVfmt to
the file instead of WAVEfmt. I still don't think I got the file size
right in the header though...